update week 3
